requirements.txt
Django==2.2.10
django-crispy-forms==1.13.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
